// src/app/types.ts
export interface CTADataProps {
  backgroundImage: string;
  title: string;
  buttonText: string;
  buttonUrl: string;
}
// ...
export interface FeatureDataProps {
  sectionTitle: string;
  buttonText: string;
  buttonUrl: string;
  image: string;
  features: {
    icon: string;
    title: string;
    description: string;
  }[];
}
// ...
export interface HeroDataProps {
  title: string;
  subtitle: string;
  btnText1: string;
  btnUrl1: string;
  btnText2: string;
  btnUrl2: string;
  funfact: {
    number: string;
    text: string;
  };
  box: {
    title: string;
    subtitle: string;
    link: string;
  };
  backgrounds: string[];
}
// ...
export interface ServiceDataProps {
  title: string;
  highlightedText: string;
  service: {
    title: string;
    subtitle: string;
    image: string;
    link: string;
    tags: { label: string; url: string }[];
    description: string;
  }[];
}
// ...
export interface TestimonialDataProps {
  sectionTitle: string;
  testimonials: {
    text: string;
    name: string;
    designation: string;
    img: string;
  }[];
}

// Video block section
export interface VideoDataProps {
  backgroundImage: string;
  videoUrl: string;
}

// Value Section Data Type
export interface ValueDataProps {
  sectionTitle: string;
  values: { text: string }[];
}

// Works Section Data Type
export interface WorksDataProps {
  title: string;
  subtitle: string;
  slides: {
    image: string;
    name: string;
    location: string;
    description: string;
  }[];
}
// Works Section Two Data Type
export interface WorksTwoDataProps {
  title: string;
  subtitle: string;
  galleryItems: GalleryDataProps[];
}

// Working Process Section Data Type
export interface WorkingProcessDataProps {
  sectionTitle: string;
  subtitle: string;
  logo: string;
  steps: {
    title: string;
    description: string;
  }[];
}